+++ VERSION_HISTORY.md
@@ -140,12 +140,16 @@
 * Counter0 mode2 routines implemented
 * Counter0 mode3 routines implemented
 * UART1 mode0 send routines implemented
-* TODO UART1 mode1 with timer2 as baud rate generator send routines implemented 
+* TODO UART1 mode1 with timer2 12T as baud rate generator send routines implemented
+    * TODO Support double baud rate
+    * TODO MCU clock frequency bias compensation.
+    * TODO Support UART1 relay and broadcast mode
+    * TODO Support UART1 interrupt mode
+
 
 Next releases todo
 * Fix delay_ms. On small ms delays it works correctly, but on seconds delay it doubled.
 * TODO Implement UART
-* Code review using AI
 * use asserts and static_assert
 * TODO Optimize timer2 baud rate calculation. Use precalculated array of baudrates
 * TODO Implement P5.4 as MCU reset source